V2: Fix wrong copying of file
